finished refactoring list code into components
