+++ README.md
@@ -4,4 +4,4 @@ This project is a work in progress.
 
 I wanted to expand on the [Scrimba Shopping List](https://github.com/BrianSlaby/scrimba-shopping-list) project that I made earlier in the [Scrimba Frontend Developer Career Path](https://scrimba.com/learn/frontend).  The idea is for this expanded project to use a Google Firebase cloud firestore database with authentication instead of the realtime database without authentication used in the original project.  I also wanted to include the option to create multiple lists to facilitate organization, likely with checkboxes (in the original project clicking on a list item simply deleted it).
 
-I also wanted to build this using React, to get more practice with it.
+I also wanted to build this using React, to get more practice with it. One focus that I've had while working on this project is to improve my file/folder structure and use of components. A majority of the project's code ended up in the Home file at first, and it was very quickly becoming difficult to navigate even with a relatively small project like this. When the time came to build the modal, I started by creating separate component files since the warning modal for deleting a list was extremely straightforward.  Then, fresh off of that work flow, I split the nested list code into separate list components.  Deciding how to structure everything was initially harder than I'd anticipated, especially since I had all of the code in front of me as a getHTML type function like I'd been using in vanilla JavaScript. Once I started splitting everything up one piece at a time and figuring out which props I'd need to pass everything clicked into place, and the resulting code was much more readable.
