--- src/components/lists/Item.jsx
@@ -1,10 +1,0 @@
-import React from "react"
-
-export default function Item({children}) {
-
-    return (
-        <li>
-            {children}
-        </li>
-    )
-}
