+++ src/components/lists/list.css
@@ -1,0 +1,83 @@
+.create-list-container,
+.create-item-container,
+.list-header-container {
+  display: flex;
+  gap: 0.5em;
+  margin: 1em 0;
+}
+
+.create-list-container {
+  margin-bottom: 1em;
+}
+
+.list-header-container {
+  background-color: var(--primary-clr);
+  border-radius: 8px;
+  display: flex;
+}
+
+.list-title {
+  text-transform: uppercase;
+  font-size: 1.125rem;
+}
+
+.list-header-btn {
+  order: -1;
+  padding: 1em;
+  background-color: transparent;
+  border: none;
+  cursor: pointer;
+  transition: transform 0.3s;
+}
+
+.list-header-btn:hover {
+  transform: scale(1.25)
+}
+
+.create-list-container > input,
+.create-item-container > input {
+  flex-grow: 1;
+  background-color: var(--lt-neutral-clr);
+}
+
+.list-delete-btn,
+.item-delete-btn {
+  cursor: pointer;
+  background-color: transparent;
+  margin: 0.5em 0.5em 0.5em auto;
+}
+
+.list-delete-btn {
+  border-radius: 8px;
+  border: 1px solid firebrick;
+  color: firebrick;
+  text-transform: uppercase;
+  transition: background-color 0.5s, color 0.5s;
+}
+
+.list-delete-btn:hover {
+  background-color: firebrick;
+  color: #FFF;
+}
+
+.item-delete-btn {
+  padding: 0.3em;
+  border: none;
+  transition: transform 0.3s;
+}
+
+.item-delete-btn:hover {
+  transform: scale(1.2);
+}
+
+.list-item-container {
+  display: flex;
+  align-items: center;
+  gap: 0.5em;
+  padding: 0.1em 0.75em;
+  border-bottom: 1px solid var(--md-neutral-clr);
+}
+
+.list-item-container:last-child {
+  border-bottom: none;
+}
